# faster assign view check [pr]