benchmarked on tiny18
```
~/code/tinygrad (master) > SCHEDULE_ONLY=1 python3 test/external/external_benchmark_schedule.py
***** model tensor in     61.82 ms
***** model schedule in  203.31 ms
all 265.20 ms
~/code/tinygrad (faster_assign_view_check) > SCHEDULE_ONLY=1 python3 test/external/external_benchmark_schedule.py
***** model tensor in     61.47 ms
***** model schedule in  198.27 ms
all 259.81 ms
```